+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 ### Added
 
+- AlphaFold export to PyMOL
+
 ### Removed
 
 ### Fixed